# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "get": {
        "operationId": "UsersController_getUsers",
        "parameters": [
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Filter users by role (e.g., LINMAS)",
            "schema": {
              "type": "string",
              "enum": [
                "KOORDINATOR",
                "LINMAS",
                "WARGA"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get list of users (filterable by role)",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_getUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user detail by ID",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update user profile (nama, HP, email, regu, alamat)",
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Nonaktifkan user (soft delete)",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/linmas": {
      "post": {
        "operationId": "UsersController_createLinmas",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLinmasDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create new Linmas member (moved from auth/register)",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/violations": {
      "post": {
        "operationId": "ViolationsController_createViolation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateViolationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Issue a Violation / SP (Koordinator Only)",
        "tags": [
          "Violations"
        ]
      },
      "get": {
        "operationId": "ViolationsController_getViolations",
        "parameters": [
          {
            "name": "linmas_id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Violation History",
        "tags": [
          "Violations"
        ]
      }
    },
...
  },
  "info": {
    "title": "MANALINMAS API",
    "description": "Manajemen Kinerja Linmas & Pelaporan Darurat API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateLinmasDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateViolationDto": {
        "type": "object",
        "properties": {
          "linmas_id": {
            "type": "string",
            "example": "uuid-of-linmas"
          },
          "violation_type": {
            "type": "string",
            "example": "Tidur saat jaga malam"
          },
          "sanction_level": {
            "type": "string",
            "enum": [
              "SP1",
              "SP2",
              "SP3"
            ],
            "example": "SP1"
          },
          "incident_date": {
            "type": "string",
            "example": "2026-05-20"
          }
        },
        "required": [
          "linmas_id",
          "violation_type",
          "sanction_level",
          "incident_date"
        ]
      },
...
    }
  }
}