# src/modules/users/dto/create-linmas.dto.ts
import {
  IsEmail,
  IsNotEmpty,
  IsOptional,
  IsString,
  IsUUID,
  MinLength,
} from 'class-validator';

export class CreateLinmasDto {
  @IsNotEmpty()
  @IsString()
  fullName!: string;

  @IsNotEmpty()
  @IsEmail()
  email!: string;

  @IsNotEmpty()
  @IsString()
  @MinLength(6)
  password!: string;

  @IsNotEmpty()
  @IsString()
  phone_number!: string;

  @IsNotEmpty()
  @IsString()
  address!: string;

  @IsOptional()
  @IsUUID()
  regu_id?: string;

  @IsOptional()
  @IsString()
  employment_date?: string; // format ISO date, akan di-parse
}